moves createStore into its own file, making it easier to write tests; stores.js now instantiates the store and is the import target for the app
